Fix .gitignore & Enable custom plugins in repo
This change enables shop owners, to add own plugins to their
repositories. Git does not allow to renable a sub path, from an already
disabled path. So this does not work:

/custom/*
!/custom/.gitkeep
!/custom/plugins/MyOwnPlugin

To enable adding own plugins to the repository, I changed the following
line:

`/custom/*`

to

`/custom/plugins/*`

Additionally I removed all superflous `!.gitkeep` statement and added a
global one at the end of the .gitignore.